// src/types/resources.ts
export type PartyPageResourcesProps = {
  party: string;
  backToHome: string;
  send: string;
  applyNowonly: string;
  forYourSpecialOccasion: string;
  thankYouForConsidering: string;
  captchaRequired: string;
  requiredName: string;
  maxLength: string;
  character: string;
  requiredLastName: string;
  requiredEmail: string;
  emailNotValid: string;
  requiredPhone: string;
  phoneValidate: string;
  requiredDate: string;
  invalidDate: string;
  lastName: string;
  canOnlyContainLettersOrComma: string;
  email: string;
  phonePlaceholder: string;
  firstName: string;
  date: string;
  cityRequired: string;
  subject: string;
  Location: string;
  theTime: string;
  requiredCity: string;
  Inquiry: string;
  requiredFirstName: string;
  subjectRequired: string;
  canOnlyContainLettersNumbersUnderscoes: string;
  subjectMinLength: string;
  requiredTime: string;
  invalidTime: string;
  requiredLocation: string;
  requiredInquiry: string;
  selectCity: string;
  invalidNumberofchickenpieces: string;
  WeddingParty: string;
  BirthdayParty: string;
  OfficeParty: string;
  SummerBackyardParty: string;
  ChristmasParty: string;
  WatchingTV: string;
  EmployeeLunch: string;
  Other: string;
  meatPreference: string;
  Numberofchickenpieces: string;
  BothWhiteDark: string;
  WhiteMeat: string;
  DarkMeat: string;
  storeRequired: string;
  selectBranch: string;
  submittedSuccess: string;
};